major: first v0.3.0 release PR (#220)
This PR lays the foundation for the next iteration of the library,
`v0.3.0`, as it introduces some breaking changes to the current API.

## Revert to single Go module

In terms of UX, the library reverts back to being a single Go module,
rather than continuing along the path of a monorepo of multiple Go
modules.

This move introduces multiple benefits, such as:

- Easier discoverability of packages within the lib, as they are
imported with the same `go get`,
- Easier experience in development and release of updated, especially
minor dependency updates.

The potential downside of this move is the introduction of more
transient dependencies for consumers; however, the Go compiler already
handles the removal of dependencies in the dependency graph that are not
used/referenced during compilation.

## Minimum version Go 1.21

The library does rely by a pretty good margin on generics, especially
when it comes to the core packages (e.g. `aggregate`, `message`, etc.)

With Go 1.21, the ergonomics around generics usage has increased
considerably, therefore we now target that specific API version.

## More functional constructors

Connected to the previous point, to increase the UX for consumers of the
library, we expect that many of the `struct` components with previously
exported fields will now have _private_ fields and require a functional
constructor, i.e. `func NewComponent()`.

The reason for this change is that generics type inference works much
better on functions than it does on direct `struct` instancing.