feat: add support for GoReleaser (#223)
Given the changes introduced in #220, the `release-please` based release
workflow is not working, and thus we need to review the whole automated
releasing process.

One tool that definitely sees more use in the Go ecosystem is
[GoReleaser](https://goreleaser.com), and since we now have a
single-module setup, we can freely use it.

This PR introduces some of the foundational steps for the new automated
release process with GoReleaser.